+++ src/persistence_internal.h
@@ -18,13 +18,14 @@ void initialise_persistent_state(void);
 /* Creates new persistent variable.  Note that type is *not* checked for
  * validity, *must* be a valid enum value! */
 void create_persistent_waveform(
-    const char *name, enum PERSISTENCE_TYPES type, size_t max_length);
+    const char *name, enum PERSISTENCE_TYPES type, unsigned int max_length);
 
 /* Updates variable from value stored on disk, returns false if no value
  * returned.  The function load_persistent_state() must be called first. */
 bool read_persistent_variable(const char *name, void *variable);
-bool read_persistent_waveform(const char *name, void *variable, size_t *length);
+bool read_persistent_waveform(
+    const char *name, void *variable, unsigned int *length);
 /* Writes value to persistent variable. */
 void write_persistent_variable(const char *name, const void *value);
 void write_persistent_waveform(
-    const char *name, const void *value, size_t length);
+    const char *name, const void *value, unsigned int length);
